requirements.txt
Django>=3.1.4
uWSGI>=2.0.18,<2.1
djangorestframework>=3.9.3
django-cors-headers>=3.6.0
psycopg2
requests